Avoid panicking in DAC stream example in favour of using stderr
This helps to avoid crashing in the case that some DACs or DAC emulators
initialise in a strange state or skip certain steps when establishing a
stream. Rather than crashing, we simply print the errors to stderr and
then continue to run. This helps to identifiy issues or strange quirks
in both physical and emulated DACs.